recommender-api/requirements.txt
aniso8601==3.0.0
atomicwrites==1.1.5
attrs==18.1.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
Cython==0.28.3
Flask==1.0.2
Flask-MySQL==1.4.0
Flask-RESTful==0.3.6
idna==2.6
itsdangerous==0.24
Jinja2==2.10
joblib==0.11
MarkupSafe==1.0
more-itertools==4.2.0
numpy==1.14.3
pandas==0.23.0
pluggy==0.6.0
py==1.5.3
PyMySQL==0.8.1
pytest==3.6.0
python-dateutil==2.7.3
pytz==2018.4
requests==2.18.4
scikit-learn==0.19.1
scikit-surprise==1.0.6
scipy==1.1.0
six==1.11.0
urllib3==1.22
Werkzeug==0.14.1
